Move wrappers for Octokit methods
