+++ src/commands/create-project.js
@@ -1,4 +1,4 @@
-const github = require("../index");
+const github = require("../lib/github");
 const printOutput = require("../lib/helpers/print-output");
 
 /**
